// src/ontologies/shacl.d.ts
import { NamedNode } from '../models.js';
export declare var _ontologyResource: NamedNode;
export declare const shacl: {
    class: NamedNode;
    datatype: NamedNode;
    declare: NamedNode;
    in: NamedNode;
    maxCount: NamedNode;
    minCount: NamedNode;
    name: NamedNode;
    node: NamedNode;
    NodeShape: NamedNode;
    optional: NamedNode;
    Parameter: NamedNode;
    PrefixDeclaration: NamedNode;
    path: NamedNode;
    prefix: NamedNode;
    property: NamedNode;
    PropertyShape: NamedNode;
    Shape: NamedNode;
    targetClass: NamedNode;
    targetNode: NamedNode;
    BlankNode: NamedNode;
    IRI: NamedNode;
    Literal: NamedNode;
    BlankNodeOrIRI: NamedNode;
    BlankNodeOrLiteral: NamedNode;
    IRIOrLiteral: NamedNode;
    nodeKind: NamedNode;
};